requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
cmake==3.25.0
contourpy==1.2.1
cycler==0.12.1
faiss-gpu==1.7.2
filelock==3.13.1
fonttools==4.53.0
fsspec==2024.6.0
huggingface-hub==0.23.3
idna==3.4
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
lit==15.0.7
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==24.1
pandas==2.2.2
pillow==10.2.0
protobuf==3.20.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.28.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentencepiece==0.2.0
six==1.16.0
sympy==1.12
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.66.4
transformers==4.41.2
triton==2.0.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.13
